requirements.txt
Django==1.9.2
django-uuslug==1.1.7
Pillow==3.1.1
python-slugify==1.2.0
Unidecode==0.4.19
